feat: Adiciona busca de estudante por id do usuario

+++ src/main/java/br/edu/ifsp/estagiei/repository/VagaRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.repository.CrudRepository;
 
 import br.edu.ifsp.estagiei.entity.Vaga;
 
-public interface VagaRepository extends CrudRepository<Vaga, Integer> {}
+public interface VagaRepository extends CrudRepository<Vaga, Long> {}
